move generated data into data directory
