+++ components/Officers/CommitteeSectionOfficers.js
@@ -1,6 +1,6 @@
 // import Image from 'next/image';
 
-import data from '../../offoutput.json';
+import data from '../../data/offoutput.json';
 import Officers from '../Officers/OfficerCard';
 
 function CommitteeOfficers({ committee }) {
