+++ pages/dev.js
@@ -5,7 +5,7 @@ import Projects from '../components/DevProjectCards';
 import Layout from '../components/Layout';
 import Officers from '../components/OfficerCard';
 import { projects } from '../data/dev';
-import data from '../offoutput.json';
+import data from '../data/offoutput.json';
 import styles from '../styles/pages/Dev.module.scss';
 
 /*const devTeamDirector = {
